move ContentHeader.vue to layout folder
