+++ src/views/InboxView/index.vue
@@ -13,7 +13,7 @@
 import StudentQuestionService from 'src/services/StudentQuestionService'
 import UserService from 'src/services/UserService'
 
-import ContentHeader from 'src/components/atoms/ContentHeader'
+import ContentHeader from 'src/components/layout/ContentHeader'
 import HelpRequests from './HelpRequests'
 
 export default {
